requirements.txt
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
transformers
accelerate
torch
-e .